Move Rails name helper to a helper module
So we don't reimplement the logic for the CLIs and Railtie separately.
