+++ lib/appsignal/cli/diagnose.rb
@@ -182,7 +182,8 @@ module Appsignal
           if rails_app?
             data[:app][:rails] = true
             current_path = Rails.root
-            initial_config[:name] = detected_rails_app_name
+            initial_config[:name] =
+              Appsignal::Utils::RailsHelper.detected_rails_app_name
             initial_config[:log_path] = current_path.join("log")
           end
