+++ lib/appsignal/cli/helpers.rb
@@ -1,18 +1,11 @@
 # frozen_string_literal: true
 
+require "appsignal/utils/rails_helper"
+
 module Appsignal
   class CLI
     module Helpers
       private
-
-      def detected_rails_app_name
-        rails_class = Rails.application.class
-        if rails_class.respond_to? :module_parent_name # Rails 6
-          rails_class.module_parent_name
-        else # Older Rails versions
-          rails_class.parent_name
-        end
-      end
 
       def ruby_2_6_or_up?
         Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
